# openapi.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "HappyPaws.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5047/"
    }
  ],
  "paths": {
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "GetCurrentUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "operationId": "UpdateCurrentUser",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "GetPublicUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicUserResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/users/me/devices": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "GetDevices",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "RegisterDevice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/devices/{id}": {
      "delete": {
        "tags": [
          "users"
        ],
        "operationId": "RemoveDevice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/v1/users/me/profile": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "GetLifestyleProfile",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifestyleProfileResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "UpsertLifestyleProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LifestyleProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LifestyleProfileResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/kyc": {
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "UploadKyc",
        "parameters": [
          {
            "name": "documentType",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DocumentType"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "document"
                ],
                "type": "object",
                "properties": {
                  "document": {
                    "$ref": "#/components/schemas/IFormFile"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KycDocumentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActivityLevel": {
        "type": "integer"
      },
...
      "BadgeResponse": {
        "required": [
          "badgeType",
          "awardedAt"
        ],
        "type": "object",
        "properties": {
          "badgeType": {
            "type": "string"
          },
          "awardedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "DeviceRequest": {
        "required": [
          "fcmToken",
          "deviceName",
          "platform"
        ],
        "type": "object",
        "properties": {
          "fcmToken": {
            "type": "string"
          },
          "deviceName": {
            "type": [
              "null",
              "string"
            ]
          },
          "platform": {
            "$ref": "#/components/schemas/Platform"
          }
        }
      },
      "DeviceResponse": {
        "required": [
          "id",
          "fcmToken",
          "deviceName",
          "platform",
          "lastActiveAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fcmToken": {
            "type": "string"
          },
          "deviceName": {
            "type": [
              "null",
              "string"
            ]
          },
          "platform": {
            "$ref": "#/components/schemas/Platform"
          },
          "lastActiveAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "DocumentStatus": {
        "type": "integer"
      },
      "DocumentType": {
        "type": "integer"
      },
...
      "HomeSize": {
        "type": "integer"
      },
      "IFormFile": {
        "type": "string",
        "format": "binary"
      },
      "KycDocumentResponse": {
        "required": [
          "id",
          "documentType",
          "status",
          "rejectionReason",
          "uploadedAt",
          "reviewedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "documentType": {
            "$ref": "#/components/schemas/DocumentType"
          },
          "status": {
            "$ref": "#/components/schemas/DocumentStatus"
          },
          "rejectionReason": {
            "type": [
              "null",
              "string"
            ]
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          },
          "reviewedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
...
      "LifestyleProfileRequest": {
        "required": [
          "homeSize",
          "activityLevel",
          "existingPetTypes",
          "hasChildren",
          "hasYard"
        ],
        "type": "object",
        "properties": {
          "homeSize": {
            "$ref": "#/components/schemas/HomeSize"
          },
          "activityLevel": {
            "$ref": "#/components/schemas/ActivityLevel"
          },
          "existingPetTypes": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "hasChildren": {
            "type": "boolean"
          },
          "hasYard": {
            "type": "boolean"
          }
        }
      },
      "LifestyleProfileResponse": {
        "required": [
          "homeSize",
          "activityLevel",
          "existingPetTypes",
          "hasChildren",
          "hasYard",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "homeSize": {
            "$ref": "#/components/schemas/HomeSize"
          },
          "activityLevel": {
            "$ref": "#/components/schemas/ActivityLevel"
          },
          "existingPetTypes": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            }
          },
          "hasChildren": {
            "type": "boolean"
          },
          "hasYard": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Platform": {
        "type": "integer"
      },
...
      "PublicUserResponse": {
        "required": [
          "id",
          "name",
          "reputationPoints",
          "badges"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "reputationPoints": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "badges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BadgeResponse"
            }
          }
        }
      },
...
      "UserProfileResponse": {
        "required": [
          "id",
          "name",
          "email",
          "avatarUrl",
          "isVerified",
          "reputationPoints",
          "badges"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "avatarUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "isVerified": {
            "type": "boolean"
          },
          "reputationPoints": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "badges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BadgeResponse"
            }
          }
        }
      }
    }
  },
...
}